code refactoring for new executor (#34970)
* code refactoring, test=develop

* refine, test=develop

* refine, test=develop

* refine, test=develop
